requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
pillow>=10.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
nltk>=3.8.1
scikit-learn>=1.3.2
numpy>=1.24.0